turn peripherals into scoped singletons
