+++ .gitignore
@@ -1,4 +1,5 @@
 *.org
 *.rs.bk
+.#*
 Cargo.lock
 target
